+++ CHANGELOG.md
@@ -36,5 +36,6 @@
 - New trait design for `LayerLifecycle` which also was renamed to `Layer`.
 - Removed low-level layer functions from `BuildContext`. They don't fit well with the design of the library at this
   point and are potential footguns. Implementing a `Layer` should work for all use-cases.
+- The `stack_id` field in `BuildContext` and `DetectContext` is now of type `StackId` instead of `String`.
 
 ## [0.3.0] 2021/09/17